Rename `_shortcodes` to `shortcodes`
