+++ apps/docs/shortcodes/client-side-navigation-hint.md
