+++ apps/docs/shortcodes/packageInstallation.js
